pyproject.toml
[project]
name = "gpuma"
version = "1.0.2"
description = "GPUMA - Geometry optimization toolkit using Fairchem UMA models and Torch-Sim"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
  { name = "Niklas Hölter", email = "niklas.hoelter@uni-muenster.de" },
]
keywords = ["chemistry", "geometry-optimization", "uma", "fairchem", "ase", "mlip", "gpuma"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Chemistry",
  "Topic :: Scientific/Engineering :: Physics",
]

dependencies = [
  "ase==3.27.0",
  "torch-sim-atomistic==0.5.1",
  "fairchem-core>=2.7,<2.10",
  "morfeus-ml>=0.7",
  "rdkit>=2022.9.5",
  "tables>=3.10.2",
  "scipy<1.15.0",
  "hf_xet",
]

[project.optional-dependencies]
yaml = ["pyyaml>=6.0"]
dev = [
    "pytest>=9.0.0",
    "build>=1.2.1",
    "twine>=5.0.0",
    "pip-tools>=7.5.2",
    "ruff>=0.14.14",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocstrings[python]>=1.0.2"
]

[project.scripts]
gpuma = "gpuma.cli:main"

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["gpuma*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E203", "D100", "D101"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D"]

[tool.ruff.format]
quote-style = "double"
line-ending = "auto"
